refactor: remove comments on self-explaining code
