+++ packages/container/src/const.ts
@@ -1,6 +1,2 @@
-/**
- *
- * Constants
- */
 export const REAGENT_WIKI_BASE = 'https://raw.githubusercontent.com/ReAGEnT-WiSe2021-22/Wiki/main/';
 export const REAGENT_WIKI_ENDPOINT = `${REAGENT_WIKI_BASE}README.md`;
