+++ packages/container/src/fixtures/party-reputation.ts
@@ -1,7 +1,3 @@
-/**
- *
- * Party Reputation mock data.
- */
 export const partyReputation = [
   {
     party: 'autem',
