+++ packages/container/src/pages/home/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactMarkdown from 'react-markdown';
-import './Home.css';
+import './index.css';
 
 /**
  *
